Refactor gem structure
Now everything is under the kolekti/metricfu directory.
