+++ features/step_definitions/metric_fu_steps.rb
@@ -1,3 +1,3 @@
 Given(/^MetricFu is registered on Kolekti$/) do
-  Kolekti.register_collector(KolektiMetricfu::Collector)
+  Kolekti.register_collector(Kolekti::Metricfu::Collector)
 end
