+++ lib/kolekti/metricfu.rb
@@ -1,0 +1,11 @@
+require 'kolekti'
+
+require "kolekti/metricfu/version"
+require "kolekti/metricfu/collector"
+require "kolekti/metricfu/parsers"
+
+module Kolekti
+  module Metricfu
+    # Your code goes here...
+  end
+end
