+++ lib/kolekti/metricfu/version.rb
@@ -1,0 +1,5 @@
+module Kolekti
+  module Metricfu
+    VERSION = "0.1.0"
+  end
+end
