--- lib/kolekti_metricfu.rb
@@ -1,9 +1,0 @@
-require 'kolekti'
-
-require "kolekti_metricfu/version"
-require "kolekti_metricfu/collector"
-require "kolekti_metricfu/parsers"
-
-module KolektiMetricfu
-  # Your code goes here...
-end
