--- lib/kolekti_metricfu/version.rb
@@ -1,3 +1,0 @@
-module KolektiMetricfu
-  VERSION = "0.1.0"
-end
