+++ spec/kolekti/metricfu/parsers/flog_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe KolektiMetricfu::Parsers::Flog do
+describe Kolekti::Metricfu::Parsers::Flog do
   it 'is expected to be a Base parser' do
-    expect(subject).to be_a(KolektiMetricfu::Parsers::Base)
+    expect(subject).to be_a(Kolekti::Metricfu::Parsers::Base)
   end
 
   describe 'class methods' do
