+++ spec/kolekti/metricfu_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe Kolekti::Metricfu do
+  it 'has a version number' do
+    expect(Kolekti::Metricfu::VERSION).not_to be nil
+  end
+end
