--- spec/kolekti_metricfu_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe KolektiMetricfu do
-  it 'has a version number' do
-    expect(KolektiMetricfu::VERSION).not_to be nil
-  end
-end
